backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==3.2.13
django-cors-headers==3.13.0
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-core==4.3.0
sqlparse==0.4.3
urllib3==1.26.13